Implement code revie and rename some properies
